Fixed the crash?
Possible explanation:
`in Entity entity` used in Move and Destroy is a `ref readonly Entity entity`, which means, once destroyed... it points to another location and therefore changes its id and world. This can mess up a lot.

Solution, caching the id before move/destroy operation and reassigning the correct archetype to the entityinfo each time.

Ran the game a couple of times and there was no crash.